# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "CoDE"
  },
  "host": "localhost",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "question"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/questions": {
      "post": {
        "tags": [
          "question"
        ],
        "summary": "Add a new question",
        "description": "",
        "operationId": "addQuestion",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Question object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Question"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/QuestionResponse"
            }
          },
          "500": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "question"
        ],
        "summary": "Update an existing question",
        "description": "",
        "operationId": "updateQuestion",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Question object that needs to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Question"
            }
          }
        ],
        "responses": {
          "203": {
            "description": "Accepted"
          },
          "500": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": [
          "question"
        ],
        "summary": "Gets all questions",
        "operationId": "getQuestion",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "202": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/QuestionResponse"
            }
          },
          "500": {
            "description": "bad request"
          }
        }
      }
    },
    "/questions/{questionId}": {
      "get": {
        "tags": [
          "question"
        ],
        "summary": "Find question by Id",
        "description": "Returns a single question",
        "operationId": "getQuestionById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "description": "Id of question to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/QuestionResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "question"
        ],
        "summary": "Delete a question by Id",
        "operationId": "deleteQuestionById",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "description": "Id of question to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "successful delete"
          },
          "500": {
            "description": "bad request"
          }
        }
      }
    }
  },
  "definitions": {
    "Question": {
      "type": "object",
      "required": [
        "title",
        "question",
        "suggestedAnswer",
        "difficulty"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "question": {
          "type": "string"
        },
        "suggestedAnswer": {
          "type": "string"
        },
        "difficulty": {
          "type": "number"
        },
        "createdBy": {
          "type": "string"
        },
        "createdDate": {
          "type": "string",
          "format": "date-time"
        },
        "modifiedBy": {
          "type": "string"
        },
        "modifiedDate": {
          "type": "string",
          "format": "date-time"
        }
      },
      "xml": {
        "name": "Question"
      }
    },
    "QuestionResponse": {
      "type": "object",
      "properties": {
        "questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Question"
          }
        }
      }
    }
  }
}